refactor: 💡 rename user files to auth for better semantics
Renamed:

BREAKING CHANGE: 🧨    - userController.ts -> authController.ts
